refactor: move DebugInfo and provider to models
