+++ pkg/compute/sensors/completed_jobs.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/bacalhau-project/bacalhau/pkg/compute/store"
-	"github.com/bacalhau-project/bacalhau/pkg/model"
+	"github.com/bacalhau-project/bacalhau/pkg/models"
 )
 
 type CompletedJobProvider struct {
@@ -16,14 +16,14 @@ func NewCompletedJobs(e store.ExecutionStore) *CompletedJobProvider {
 }
 
 // GetDebugInfo implements model.DebugInfoProvider
-func (c *CompletedJobProvider) GetDebugInfo(ctx context.Context) (model.DebugInfo, error) {
+func (c *CompletedJobProvider) GetDebugInfo(ctx context.Context) (models.DebugInfo, error) {
 	jobcounts, err := c.ExecutionStore.GetExecutionCount(ctx, store.ExecutionStateCompleted)
-	return model.DebugInfo{
+	return models.DebugInfo{
 		Component: "jobsCompleted",
 		Info:      jobcounts,
 	}, err
 }
 
-var _ model.DebugInfoProvider = (*CompletedJobProvider)(nil)
+var _ models.DebugInfoProvider = (*CompletedJobProvider)(nil)
 
 // add a method to LocalExecutionState store interface to return an execution count for a given state.
