+++ pkg/orchestrator/selection/discovery/info_provider.go
@@ -3,7 +3,7 @@ package discovery
 import (
 	"context"
 
-	"github.com/bacalhau-project/bacalhau/pkg/model"
+	"github.com/bacalhau-project/bacalhau/pkg/models"
 	"github.com/bacalhau-project/bacalhau/pkg/orchestrator"
 )
 
@@ -11,16 +11,16 @@ type discoveredNodesProvider struct {
 	discoverer orchestrator.NodeDiscoverer
 }
 
-func NewDebugInfoProvider(discoverer orchestrator.NodeDiscoverer) model.DebugInfoProvider {
+func NewDebugInfoProvider(discoverer orchestrator.NodeDiscoverer) models.DebugInfoProvider {
 	return &discoveredNodesProvider{discoverer: discoverer}
 }
 
 // GetDebugInfo implements models.DebugInfoProvider
-func (p *discoveredNodesProvider) GetDebugInfo(ctx context.Context) (info model.DebugInfo, err error) {
+func (p *discoveredNodesProvider) GetDebugInfo(ctx context.Context) (info models.DebugInfo, err error) {
 	nodes, err := p.discoverer.List(ctx)
 	info.Component = "DiscoveredNodes"
 	info.Info = nodes
 	return info, err
 }
 
-var _ model.DebugInfoProvider = (*discoveredNodesProvider)(nil)
+var _ models.DebugInfoProvider = (*discoveredNodesProvider)(nil)
